requirements-minimal.txt
# Minimal requirements for CI/CD - fallback if main requirements.txt fails
fastapi>=0.100.0
uvicorn>=0.20.0
alpaca-py>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.0.0
python-multipart>=0.0.6
requests>=2.28.0
pandas>=2.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
httpx>=0.24.0
loguru>=0.7.0
redis>=5.0.0
PyYAML>=6.0.0
cryptography>=40.0.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
PyJWT>=2.6.0